Backend/Fitness_chatbot/fitness.py
from phi.agent import Agent
from phi.model.groq import Groq
from phi.tools.wikipedia import WikipediaTools
from phi.tools.duckduckgo import DuckDuckGo

# loading environment variables from a .env file into our Python application's environmen
from dotenv import load_dotenv
load_dotenv()



fitness_agent = Agent(
    name="Fitness AI Agent",
    model=Groq(id="llama-3.2-1b-preview"),
    tools=[
        WikipediaTools(),
        DuckDuckGo()  # Add search capability to a single agent
    ],
    instructions=[
        "You are an expert fitness trainer and nutrition specialist with extensive knowledge in exercise science, nutrition, and wellness.",
        "Create personalized fitness and nutrition plans based on user's goals, needs, and limitations.",
        "Always present workout plans, meal plans, and schedules in well-structured tables.",
        "Ask for key information if not provided: age, weight, height, fitness level, goals, limitations/injuries, dietary preferences, and available equipment.",
        "Provide exercise recommendations with proper form instructions, sets, reps, and rest periods.",
        "Calculate personalized caloric and macronutrient needs based on user information.",
        "Design meal plans that align with user's fitness goals and dietary preferences.",
        "Include both workout and recovery strategies in your recommendations.",
        "Use markdown formatting to enhance readability of your responses.",
        "Cite reputable sources for any specific claims or recommendations.",
        "Prioritize user safety and recommend professional consultation when appropriate.",
        "Maintain an encouraging, positive, and motivational tone in all interactions.",
        "Track user progress and adjust recommendations based on feedback.",
        "Include measurement units in both metric and imperial when relevant."
    ],
    show_tool_calls=True,
    markdown=True,
)

# Then use just this agent directly
fitness_agent.print_response(
    "I am 30 year old i want a personlaised nutrition plan for 4 days in a week ,i am working hard to loose weight",
    stream=True  # Enable streaming for faster initial response
)