requirements.txt
TIME-python
pyttsx3==2.90
opencv-contrib-python==4.7.0.72
dotenv
Flask-Cors
fastapi
uvicorn
groq
co
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
wikipedia
mysql-connector-python
opencv-python
mediapipe
numpy
